EAH - add typescript to a couple of ui/public/utils files (#23946)
* typescriptin a few ui/public

* remove mistaken commits

* typescript numeric

* ts migrateLegacyQuery

* object or string type

* no need for file extension on import statement

* revert mistaken commit

* type fix - only object

* snake case!
