+++ src/ui/public/utils/migrate_legacy_query.ts
@@ -17,14 +17,19 @@
  * under the License.
  */
 
-export class StringUtils {
+import { has } from 'lodash';
 
-  /**
-   * Returns a version of the string with the first letter capitalized.
-   * @param str {string}
-   * @returns {string}
-   */
-  static upperFirst(str) {
-    return str ? str.charAt(0).toUpperCase() + str.slice(1) : '';
+/**
+ * Creates a standardized query object from old queries that were either strings or pure ES query DSL
+ *
+ * @param query - a legacy query, what used to be stored in SearchSource's query property
+ * @return Object
+ */
+export function migrateLegacyQuery(query: object): object {
+  // Lucene was the only option before, so language-less queries are all lucene
+  if (!has(query, 'language')) {
+    return { query, language: 'lucene' };
   }
+
+  return query;
 }
