+++ src/ui/public/utils/numeric.ts
@@ -17,15 +17,8 @@
  * under the License.
  */
 
-import { partition } from 'lodash';
+import _ from 'lodash';
 
-export function sortPrefixFirst(array, prefix, property) {
-  if (!prefix) return array;
-  const lowerCasePrefix = ('' + prefix).toLowerCase();
-
-  const partitions = partition(array, entry => {
-    const value = ('' + (property ? entry[property] : entry)).toLowerCase();
-    return value.startsWith(lowerCasePrefix);
-  });
-  return [ ...partitions[0], ...partitions[1] ];
+export function isNumeric(v: any): boolean {
+  return !_.isNaN(v) && (typeof v === 'number' || (!Array.isArray(v) && !_.isNaN(parseFloat(v))));
 }
