+++ src/ui/public/utils/sort_prefix_first.ts
@@ -17,8 +17,17 @@
  * under the License.
  */
 
-import _ from 'lodash';
+import { partition } from 'lodash';
 
-export function isNumeric(v) {
-  return !_.isNaN(v) && (typeof v === 'number' || (!Array.isArray(v) && !_.isNaN(parseFloat(v))));
+export function sortPrefixFirst(array: any[], prefix?: string | number, property?: string): any[] {
+  if (!prefix) {
+    return array;
+  }
+  const lowerCasePrefix = ('' + prefix).toLowerCase();
+
+  const partitions = partition(array, entry => {
+    const value = ('' + (property ? entry[property] : entry)).toLowerCase();
+    return value.startsWith(lowerCasePrefix);
+  });
+  return [...partitions[0], ...partitions[1]];
 }
