+++ src/ui/public/utils/string_utils.ts
@@ -17,20 +17,13 @@
  * under the License.
  */
 
-import { has } from 'lodash';
-
-/**
- * Creates a standardized query object from old queries that were either strings or pure ES query DSL
- *
- * @param query - a legacy query, what used to be stored in SearchSource's query property
- * @return Object
- */
-export function migrateLegacyQuery(query) {
-
-  // Lucene was the only option before, so language-less queries are all lucene
-  if (!has(query, 'language')) {
-    return { query: query, language: 'lucene' };
+export class StringUtils {
+  /**
+   * Returns a version of the string with the first letter capitalized.
+   * @param str {string}
+   * @returns {string}
+   */
+  public static upperFirst(str: string): string {
+    return str ? str.charAt(0).toUpperCase() + str.slice(1) : '';
   }
-
-  return query;
 }
